+++ examples/read_all_dlpc2607_reg.cc
@@ -7,7 +7,7 @@
 
 int main() {
   I2cBus i2c_1(1);
-  int addr = 0x1B;  // DLPC2607 address
+  uint16_t addr = 0x1B;  // DLPC2607 address
   uint8_t value_out[4] = {0};
   int reg[256] = {0};  // DLP 32-bit register values
